requirements.txt
Flask==3.1.2
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
injector==0.22.0
openai==2.7.0
pytest==8.4.2
python-dotenv==1.2.1
WTForms==3.2.1
